requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0     
tzdata==2025.2
urllib3==2.6.1